chore(examples): generate index.html with npm `build` script
