+++ .gitignore
@@ -2,4 +2,3 @@ bower_components
 node_modules
 npm-debug.log
 dist
-index.html
