+++ examples/webpack.config.js
@@ -1,8 +1,27 @@
 const path = require('path');
 const bemLoader = require('bem-loader');
 const CollectBemAssetsPlugin = bemLoader.CollectBemAssetsPlugin;
+
 const fs = require('fs-extra');
 const bemxjst = require('bem-xjst');
+const bemhtml = bemxjst.bemhtml;
+
+const templates = bemhtml.compile([
+    fs.readFileSync('./node_modules/bem-core/common.blocks/ua/ua.bemhtml'),
+    fs.readFileSync('./node_modules/bem-core/common.blocks/page/page.bemhtml'),
+    fs.readFileSync('./node_modules/bem-core/common.blocks/page/__js/page__js.bemhtml'),
+    fs.readFileSync('./node_modules/bem-core/common.blocks/page/__css/page__css.bemhtml'),
+    fs.readFileSync('./components/react/react.bemhtml.js')
+].join(''));
+
+fs.outputFile(
+    './dist/index.html',
+    templates.apply({
+        block: 'page',
+        scripts: { elem: 'js', url: './main.js' },
+        content: { block: 'react' }
+    })
+);
 
 module.exports = {
     entry: './index.js',
